feat(yellow): add Pokémon Yellow game data and support
- Add Yellow game catalog, rules provider, maps, encounters, Pokédex, and world data
- Display encounter probabilities with decimal precision

+++ PokemonFieldGuide/Models/FieldGuideData.cs
@@ -38,7 +38,7 @@ public sealed class Encounter
     public string SpeciesId { get; set; } = "";
     public int MinLevel { get; set; }
     public int MaxLevel { get; set; }
-    public int Chance { get; set; }
+    public double Chance { get; set; }
     public string Method { get; set; } = "";
     public string Version { get; set; } = "Both";
 }
